bulk account loader freq is float

+++ src/driftpy/accounts/bulk_account_loader.py
@@ -29,7 +29,7 @@ class BulkAccountLoader:
         self,
         connection: AsyncClient,
         commitment: Commitment = "confirmed",
-        frequency: int = 1,
+        frequency: float = 1,
     ):
         self.connection = connection
         self.commitment = commitment
